Normalizes most of the function names.
I wrote a script that parsed all of the files and normalized the function names. there are some that are "special" and will need to be manually looked over.

I would recommend creating some kind og a list of function names the user can pick from. I know there will be special use cases and there should also be a means to input the custom function name. When a custom name is entered the code should be flagged for review. This way you have some consistency. You will also be able to provide the user with a fantastic search. they can select the manufacturer of the remote and then select the function name they want from the list. This will retrieve all codes associated with the matching function name. then the user can decide which ones they want to use.

I know using the device id's and sub id's is easier for you and I. but the reality of it is the user is not going to know what in the world those are.. for example if wanting to lookup the Volume+ for a Sony AVR they are most likely not going to know the device id can be any one of the following ids 0x030, 0x00D, 0x02D, 0x00C, 0x02C, 0x012, 0x018, 0x090, 0x0B0 so it is better to enumerate all of the functions that match the device type that is selected. you can also do an explicit model lookup if this information was stored somewhere. I did not stumble across this in the database files as of yet if it does exist.

But the trick is to control how users are naming the functions so they will be the same across the board. almost all brands of remotes share the same button and same button names.. the other thing I noticed is that on many occasions a remote will have buttons on it that are not associated to the device type of the actual device.. We will use Sony again.. There can be buttons for controlling a TV on an AVR remote, these buttons use the device ID that matches up with the TV's A user is going to name the button TV_SOME_FUNCTION_NAME.. The function name should give the user a clue that the code that is used is not an AVR code but it is a TV code. This would need to be done in the function name. I have not come up with a nomenclature for this I am sure you could come up with something.

I am doing this house cleaning to get everything primed to build a mysql database with the information.

If you have a list of device id's to friendly names I would love to be able to also add this to the database.
The manufacturer names are already there. and if there are no model numbers I am going to set the database up so that model numbers can be used if wanted..