Add Peekton PK 1655 TNT (4,251.csv) to database
I essentially used the JP1 file, converted it with RemoteManager to GIRR, and finally converted it to IRDB with IrScrutinizer. 

Side-note: all these efforts were useless because at the end I couldn't use the Peekton for the purpose I intended. If anyone cares, no, you cannot record the VHS input to the USB storage. But it was a great experience to learn about this world, tho.